requirements.txt

asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
channels==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
kombu==5.3.2
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3.post1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
